Remove the attempt to run as non-privileged user
This makes sense when running a production service, but
we're using our CI pipeline - so the risk is only that the
CI step (which has very few permissions) is compromised.

The reason I'm reaching to remove it is because using a new
user requires chown-ing the working directory, but this
pipeline is set up to mount the working directory as a volume,
so any chmod/chown changes affect the agent.